Revert "IR-3575 playmode on scene change (#25)"
This reverts commit 6cf598661d9b92474222ee8737ef4a524a6dec4e.
